matrix buildx by tool, support cache of snort/daq as download cannot be made reliable.

+++ snort/downloadhelper.sh
@@ -9,14 +9,23 @@ tarfile=$2
 retries=5
 if [[ "$url" == "" || "$tarfile" == "" ]] ; then
 	echo need URL and tarfile
+	exit 1
+fi
+
+if [[ -f "$tarfile" ]] ; then
+	echo $tarfile exists, skipping download
+	exit 0
 fi
 
 i=0
 while [ $i -lt $retries ]; do
 	i=$((i+1))
 	rm -f $outfile
+	# TODO: workaround curl segfault in getaddrinfo() handling redirect under qemu
+	finalurl=$(curl -Ls -w %{url_effective} -o /dev/null $url)
+	echo final URL: $finalurl
 	# TODO: snort binary serving does not work with TLS 1.3
-	curl -Lv $url --tlsv1.2 --tls-max 1.2 --output $tarfile --trace -
+	curl -v "$finalurl" --tlsv1.2 --tls-max 1.2 --output $tarfile --trace -
 	tar ztvf $tarfile
 	tarstatus=$?
 	if [[ -f "$tarfile" && $tarstatus -eq 0 ]] ; then
@@ -25,7 +34,7 @@ while [ $i -lt $retries ]; do
 		break
 	fi
 	echo retrying....
-	sleep 5
+	sleep 60
 done
 
 echo failed to download $tarfile.
